Add from_axis_angle_deg() and fix sign
Matrix44.from_axis_angle() was assuming radians for the input angle, but that
is different than the convention of using degrees in thes module.

Add from_axis_angle_deg() and from_axis_angle_rad().
Might make a from_axis_angle the defaults to degrees, but will wait for
now to so it's obvious if that would break anything.

Also, flip the signs (same as switching the matrix index order).
The resulting signs for x, y, and z incorrectly negative until this change.

Also add minimal tests for from_axis_angle_deg()